* encapsulated random number handling functions into a static Random class and   removed them from globals.h * removed executable flags from some files (no idea how they get set) * fixed floats in planar graph member definition
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@17551 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzwgen/inc/GraphNode.h
@@ -56,7 +56,7 @@ public:
       an empty list of connections.*/
   GraphNode() {} 
   /** Constructor which takes Point2D data as parameters. */
-  GraphNode(double x, double y) : Vector2Df(x,y) {} 
+  GraphNode(float x, float y) : Vector2Df(x,y) {} 
   /** Connects the current node to the one passed. Adds the passed node to the 
       connections list for both nodes. */
   void connect(GraphNodePtr node) {
